feat(all): dropped createTranspiler, updated examples and changed tab size
createTranspiler has been dropped in plugins in favour of having that logic inside core. This way,
it helps the plugin to reduce in size and there is only 1 source of truth. All of the plugins have
been updated to reflect that change and so have the examples. Tab size has also been changed from 4
to 2 to align with main JS standards.

BREAKING CHANGE: createTranspiler is no longer available.